Safari-611.1.10/LayoutTests/accessibility/ios-simulator/iframe-access-expected.txt


This makes sure that an iframe is accessible on iOS

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXIdentifier: content
AXIdentifier: iframe
AXIdentifier:
AXIdentifier:
AXIdentifier:
AXIdentifier: frame-button
AXIdentifier: frame-link
AXIdentifier:
PASS successfullyParsed is true

TEST COMPLETE
